+++ mex/extractors/sorters.py
@@ -15,7 +15,9 @@ def topological_sort(
 ) -> None:
     """Sort the given list of items in-place according to their topology.
 
-    Items can refer to each other using key fields. A parent item can reference a child item by storing the child’s `primary_key` in the parent’s `child_key` field. Similarly, a child can reference its parent using the `parent_key` field.
+    Items can refer to each other using key fields. A parent item can reference a child
+    item by storing the child's `primary_key` in the parent's `child_key` field.
+    Similarly, a child can reference its parent using the `parent_key` field.
 
     This can be useful for submitting items to the backend in the correct order.
     """