+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - add `topological_sort` function to reorder a list of items based on topology
+- transform endnote
 
 ### Changes